+++ src/index.js
@@ -6,20 +6,6 @@ import AlpineScreenMagicMethod from './screen'
 import AlpineScrollMagicMethod from './scroll'
 import AlpineTruncateMagicMethod from './truncate'
 
-const alpine = window.deferLoadingAlpine || ((alpine) => alpine())
-
-window.deferLoadingAlpine = function (callback) {
-    AlpineComponentMagicMethod.start()
-    AlpineFetchMagicMethod.start()
-    AlpineIntervalMagicMethod.start()
-    AlpineRangeMagicMethod.start()
-    AlpineScreenMagicMethod.start()
-    AlpineScrollMagicMethod.start()
-    AlpineTruncateMagicMethod.start()
-
-    alpine(callback)
-}
-
 export default {
     AlpineComponentMagicMethod,
     AlpineFetchMagicMethod,