+++ tests/screen.spec.js
@@ -3,7 +3,6 @@ import AlpineScreenMagicMethod from '../dist/screen'
 import { waitFor } from '@testing-library/dom'
 
 beforeAll(() => {
-    window.ontouchstart = jest.fn()
     window.Alpine = Alpine
 })
 
@@ -51,37 +50,4 @@ test('$screen > breakpoint could be provided as number', async () => {
     await waitFor(() => {
         expect(document.querySelector('span').getAttribute('style')).toEqual('display: none;')
     })
-})
-
-test('$screen > can detect orientation', async () => {
-    document.body.innerHTML = `
-      <div x-data>
-          <span x-show="$screen('landscape')"></span>
-      </div>
-  `
-
-    Alpine.start()
-
-    expect(document.querySelector('span').getAttribute('style')).toEqual(null)
-
-    window.innerHeight = 2000
-
-    window.dispatchEvent(new Event('resize'))
-
-    await waitFor(() => {
-        expect(document.querySelector('span').getAttribute('style')).toEqual('display: none;')
-    })
-})
-
-test('$screen > can detect touch screens', async () => {
-    document.body.innerHTML = `
-      <div x-data>
-          <p x-show="$screen('touch')"></p>
-          <span x-show="!$screen('touch')"></span>
-      </div>
-  `
-
-    Alpine.start()
-    expect(document.querySelector('p').getAttribute('style')).toEqual(null)
-    expect(document.querySelector('span').getAttribute('style')).toEqual('display: none;')
 })