+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Helpers were registered twice by mistake resulting in unecessary function calls.
+
+### Added
+- Added config objets to allow users to customise some helpers
+
+## [0.4.1] - 2020-11-16
+### Fixed
+- [$truncate]: Fixes an issue where the ellipsis would show even if the source length was shorter.
 
 ## [0.4.0] - 2020-11-15
 ### Added